+++ src/lib.rs
@@ -1,8 +1,9 @@
 pub mod matrix;
-pub mod vector;
+#[macro_use]
+pub mod vector_macro;
 
 #[allow(unused_imports)]
 pub use matrix::*;
 
 #[allow(unused_imports)]
-pub use vector::*;
+pub use vector_macro::*;